g needs to learn git (firsts commits to get to the edited website
