+++ src/pages/test.tsx
@@ -1,0 +1,8 @@
+import Footer from "../components/footer";
+import Header2 from "../components/header";
+import Image from "next/image";
+import Link from "next/link";
+
+const Test = () => <div className=""></div>;
+
+export default Test;
